Split some mixin classes out into their own files
- mixins.py is becoming quite bloated!
